auth/requirements.txt
flask
gunicorn
flask-cors
asyncpg
python-dotenv
requests
